requirements.txt
# ── HTTP & Retry ───────────────────────────────────────────────────────────────
httpx==0.27.0
requests==2.31.0
tenacity==8.2.3

# ── Market Data ────────────────────────────────────────────────────────────────
yfinance==0.2.66
feedparser==6.0.11

# ── Configuration & Validation ─────────────────────────────────────────────────
pyyaml==6.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1

# ── Telegram ───────────────────────────────────────────────────────────────────
python-telegram-bot==21.3

# ── Technical Analysis ─────────────────────────────────────────────────────────
pandas==2.2.1
numpy==1.26.4

# ── Scheduling (APScheduler fallback) ──────────────────────────────────────────
APScheduler==3.10.4

# ── GCP Secret Manager (optional) ─────────────────────────────────────────────
google-cloud-secret-manager==2.20.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
responses==0.25.0
